Switch from argparse to click
