+++ docker/pyproject.toml
@@ -10,12 +10,13 @@ colorama = "^0.4.3"
 tabulate = "^0.8.6"
 progressbar2 = "^3.47.0"
 aiohttp = "^3.6.2"
+click = "^7.0"
 
 [tool.poetry.dev-dependencies]
 
+[tool.poetry.scripts]
+ghd = 'ghd:run_main'
 [build-system]
 requires = ["poetry>=1.0.0"]
 build-backend = "poetry.masonry.api"
 
-[tool.poetry.scripts]
-ghd = 'ghd:run_main'
